opencode_django/requirements.txt
# Django and core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-ninja>=1.0,<2.0
django-cors-headers>=4.0,<5.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Database
psycopg[binary]>=3.2,<4.0
psycopg-pool>=3.2,<4.0

# WebSocket
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0

# Stripe
stripe>=8.0,<9.0

# Utilities
python-dateutil>=2.8,<3.0
uuid>=1.30,<2.0

# ASGI server
uvicorn>=0.25,<1.0

# Production
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
